MdeModulePkg: Duplicate BaseRngLibTimerLib to MdeModulePkg
The BaseRngLibTimerLib allows to generate number based on a timer.
This mechanism allows to have a basic non-secure implementation
for non-production platforms.
To bind and identify Random Number Generators implementations with
a GUID, an unsafe GUID should be added. This GUID cannot be added
to the MdePkg unless it is also added to a specification.

To keep the MdePkg self-contained, copy the BaseRngLibTimerLib to
the MdeModulePkg. This will allow to define an unsafe Rng GUID
in a later patch in the MdeModulePkg.

The MdePkg implementation will be removed later. This allows to give
some time to platform owners to switch to the MdeModulePkg
implementation.

Signed-off-by: Pierre Gondois <pierre.gondois@arm.com>

# Conflicts:
#	ArmVirtPkg/ArmVirt.dsc.inc
#	EmulatorPkg/EmulatorPkg.dsc
#	MdeModulePkg/MdeModulePkg.dsc
#	OvmfPkg/AmdSev/AmdSevX64.dsc
#	OvmfPkg/Bhyve/BhyveX64.dsc
#	OvmfPkg/CloudHv/CloudHvX64.dsc
#	OvmfPkg/IntelTdx/IntelTdxX64.dsc
#	OvmfPkg/Microvm/MicrovmX64.dsc
#	OvmfPkg/OvmfPkgIa32.dsc
#	OvmfPkg/OvmfPkgIa32X64.dsc
#	OvmfPkg/OvmfPkgX64.dsc
#	OvmfPkg/OvmfXen.dsc
#	OvmfPkg/RiscVVirt/RiscVVirt.dsc.inc
#	SecurityPkg/SecurityPkg.dsc
#	SignedCapsulePkg/SignedCapsulePkg.dsc